ENH: get openfoam case in folder
